6,7 uploaded vers

+++ Petros/set6/46/symtab/insert.cc
@@ -1,17 +1,17 @@
 #include "symtab.ih"
 
-void Symtab::insert(string const &nextSym, size_t idx)
+void Symtab::insert(string const &nextSym, int idx)
 {
     if (d_size == d_capacity)
         enlarge(d_capacity * 2);
 
-    for (size_t idxCopy = d_size; idxCopy != 0; idxCopy--)
+    for (int idxCopy = d_size; idxCopy != 0; idxCopy--)
         if (idxCopy > idx)
         {
             d_symbols[idxCopy] = d_symbols[idxCopy - 1];
-            d_symbols[idxCopy]->assign(Value(int(idxCopy)));
+            d_symbols[idxCopy]->assign(Value(idxCopy));
         }
 
-    d_symbols[idx] = new Symbol(nextSym, int(idx));
+    d_symbols[idx] = new Symbol(nextSym, idx);
     ++d_size;
 }
